+++ kernel/include/klib.h
@@ -105,7 +105,7 @@ void	act_init(boot_info_t *bi);
 void	act_wait(int act, aid_t next_hint);
 void *	act_register(const reg_frame_t * frame, const char * name);
 void *	act_get_ref(act_t * ctrl);
-uint64_t	act_get_id(act_t * ctrl);
+size_t	act_get_id(act_t * ctrl);
 int	act_get_status(act_t * ctrl);
 int	act_revoke(act_t * ctrl);
 int	act_terminate(act_t * ctrl);
